+++ tests/Linna/DataMapper/Uuid4Test.php
@@ -11,14 +11,14 @@ declare(strict_types=1);
 
 namespace Linna\DataMapper;
 
-//use Linna\DataMapper\UUID4;
+//use Linna\DataMapper\Uuid4;
 use PHPUnit\Framework\TestCase;
 use InvalidArgumentException;
 
 /**
  * Universally Unique Identifier Version 4 utility Test
  */
-class UUIDTest extends TestCase
+class Uuid4Test extends TestCase
 {
     /**
      * Test new object instance.
@@ -27,43 +27,43 @@ class UUIDTest extends TestCase
      */
     public function testNewObjectInstance(): void
     {
-        $this->assertInstanceOf(UUID4::class, new UUID4());
-        $this->assertInstanceOf(UUID4::class, new UUID4('691ce06d-0afb-428d-b5fc-5384cd3046fb'));
+        $this->assertInstanceOf(Uuid4::class, new Uuid4());
+        $this->assertInstanceOf(Uuid4::class, new Uuid4('691ce06d-0afb-428d-b5fc-5384cd3046fb'));
     }
 
     /**
-     * Test new object instance with worng UUID.
+     * Test new object instance with worng Uuid.
      *
      * @return void
      */
-    public function testNewObjectInstanceWithWrongUUID(): void
+    public function testNewObjectInstanceWithWrongUuid(): void
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('Invalid UUID version 4 provided.');
 
-        (new UUID4('691ce06d-0afb-428d-b5fc'));
+        (new Uuid4('691ce06d-0afb-428d-b5fc'));
     }
 
     /**
-     * Test UUID with get hex.
+     * Test Uuid4 with get hex.
      *
      * @return void
      */
-    public function testUUIDWithGetHex(): void
+    public function testUuid4WithGetHex(): void
     {
-        $uuid = new UUID4('691ce06d-0afb-428d-b5fc-5384cd3046fb');
+        $uuid = new Uuid4('691ce06d-0afb-428d-b5fc-5384cd3046fb');
 
         $this->assertSame('691ce06d-0afb-428d-b5fc-5384cd3046fb', $uuid->getHex());
     }
 
     /**
-     * Test UUID with get bin.
+     * Test Uuid4 with get bin.
      *
      * @return void
      */
-    public function testUUIDWithGetBin(): void
+    public function testUuid4WithGetBin(): void
     {
-        $uuid = new UUID4('691ce06d-0afb-428d-b5fc-5384cd3046fb');
+        $uuid = new Uuid4('691ce06d-0afb-428d-b5fc-5384cd3046fb');
 
         $this->assertSame(\hex2bin(\str_replace('-', '', '691ce06d-0afb-428d-b5fc-5384cd3046fb')), $uuid->getBin());
     }
@@ -75,7 +75,7 @@ class UUIDTest extends TestCase
      */
     public function testGenerateWithGetHex(): void
     {
-        $uuid = new UUID4();
+        $uuid = new Uuid4();
         $regex = '/^[0-9a-f]{8}-[0-9a-f]{4}-[4][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i';
 
         $this->assertMatchesRegularExpression($regex, $uuid->getHex());
@@ -89,7 +89,7 @@ class UUIDTest extends TestCase
      */
     public function testGenerateWithGetBin(): void
     {
-        $uuid = new UUID4();
+        $uuid = new Uuid4();
 
         $this->assertSame(\hex2bin(\str_replace('-', '', $uuid->getHex())), $uuid->getBin());
     }